Move registry logic into its own place
This is so we can add hostfxr logic to find the path in the registry and dedupe logic and isolate the components.
